move calculateTripDuration function to utils folder for better reusability
